+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Changed
 - install command line interface via `setuptools.setup(entry_points=…)`
+- now private:
+  - `tooncher._LoginDelayed`
+  - `tooncher._LoginSuccessful`
 
 ### Removed
 - `argcomplete`